Now runs on Windows. New version of System.Data.SQLite.DLL, 1.0.103.0 with corresponding SQLite.Interop.dll in folder where the exe file is.

+++ Updater.cs
@@ -489,7 +489,7 @@ namespace Ketarin
                     catch (SQLiteException ex)
                     {
                         // If "locked" exception (slow USB device eg.) continue trying
-                        if (ex.ErrorCode == SQLiteErrorCode.Locked)
+                        if (ex.ErrorCode == (int)SQLiteErrorCode.Locked)
                         {
                             numTries--;
                             LogDialog.Log(job, ex);
